Filter for a cell from which to select others

anchor protects you from selecting initial cells that do not exist. This is useful when choosing inital cells from which to search for others, using offset and extend. If the chosen cells do not exist, then cells are generated at the given rows and columns, with NAs in every column other than 'row' and 'col'.

anchor(cells, rows, cols, cross = TRUE)

Arguments

cells

Data frame, the cells among which the anchor will be sought

rows

Numeric, the rows of the intended cells

cols

Numeric, the columns of the intended cell

cross

Logical, whether to 'cross' the given rows and columns to generate all combinations. Unless this is TRUE, rows and cols must be the same length. TRUE by default.

Examples

cells <- tidytable(purpose$`NNW WNW`)
## Not run: ------------------------------------
#   anchor(cells, 0, 1)
#   anchor(cells, 1, 1:2)
## ---------------------------------------------
anchor(cells, 1, 1)

#> # A tibble: 1 × 6
#>     row   col character double integer logical
#>   <int> <int>     <chr>  <dbl>   <int>   <lgl>
#> 1     1     1        X2     NA      NA      NA

anchor(cells, 1, 1:2, cross = TRUE)

#> # A tibble: 2 × 6
#>     row   col character double integer logical
#>   <int> <int>     <chr>  <dbl>   <int>   <lgl>
#> 1     1     1        X2     NA      NA      NA
#> 2     1     2        X3     NA      NA      NA

anchor(cells, 100, 1)

#> # A tibble: 1 × 6
#>     row   col character double integer logical
#>   <int> <int>     <chr>  <dbl>   <int>   <lgl>
#> 1   100     1      <NA>     NA      NA      NA