Extend a bag of cells in one direction, optionally up to a boundary condition.

A bag of data cells is a data frame with at least the columns 'row' and 'col', as well as any others that carry information about the cells, e.g. their values. More cells may be added to the bag by extending the bag in a given direction, either by a number of rows or columns, or up to (and optionally including) cells that meet a boundary condition. The boundary may be required to be detected in every cell along the 'leading edge' of the bag, otherwise extension will stop at the nearest boundary that is detected.

extend(bag, cells, direction, n = NULL, boundary = NULL, edge = FALSE,
  include = FALSE)

extend_N(bag, cells, n = NULL, boundary = NULL, edge = FALSE,
  include = FALSE)

extend_E(bag, cells, n = NULL, boundary = NULL, edge = FALSE,
  include = FALSE)

extend_S(bag, cells, n = NULL, boundary = NULL, edge = FALSE,
  include = FALSE)

extend_W(bag, cells, n = NULL, boundary = NULL, edge = FALSE,
  include = FALSE)

Arguments

bag
Data frame. The original selection, including at least the columns 'row' and 'column', which are numeric/integer vectors.

cells
Data frame. All the cells in the sheet, among which to extend the bag (extensions beyond existing cells will be padded with blank cells). Must include at least the columns 'row' and 'column', as well as any columns referred to by the boundary formula.

direction
Character vector length 1. The direction in which to extend, among the compass directions "N", "E", "S", "W", where "N" is north (up).

n
Integer vector length 1, >= 0. The number of rows/cols to extend by in the given direction.

boundary
Formula to express a boundary condition. `~ col <= 50` would go up to the 50th column. NAs are treated the same as FALSE, but with a warning.

edge
Logical vector length 1. Whether to require the boundary formula to be TRUE along the entire leading edge of the bag that is being extended.

include
Logical vector length 1. Whether to include in the extension the first cell (and its row/col of fellow cells) at which the boundary condition is met.

Details

A bag may have ragged rows or ragged cols. Gaps will be filled in, even when n = 0.

Functions

extend_N: Extend a bag of cells to the north

extend_E: Extend a bag of cells to the east

extend_S: Extend a bag of cells to the south

extend_W: Extend a bag of cells to the west

Examples

# Load some pivoted data
(x <- purpose$`NNW WNW`)

#>                            X2      X3     X4     X5    X6     X7
#> 1                        <NA>    <NA> Female   <NA>  Male   <NA>
#> 2                        <NA>    <NA>  0 - 6 7 - 10 0 - 6 7 - 10
#> 3           Bachelor's degree 15 - 24   7000  27000  <NA>  13000
#> 4                        <NA> 25 - 44  12000 137000  9000  81000
#> 5                        <NA> 45 - 64  10000  64000  7000  66000
#> 6                        <NA>     65+   <NA>  18000  7000  17000
#> 7                 Certificate 15 - 24  29000 161000 30000 190000
#> 8                        <NA> 25 - 44  34000 179000 31000 219000
#> 9                        <NA> 45 - 64  30000 210000 23000 199000
#> 10                       <NA>     65+  12000  77000  8000 107000
#> 11                    Diploma 15 - 24   <NA>  14000  9000  11000
#> 12                       <NA> 25 - 44  10000  66000  8000  47000
#> 13                       <NA> 45 - 64   6000  68000  5000  58000
#> 14                       <NA>     65+   5000  41000  1000  34000
#> 15           No Qualification 15 - 24  10000  43000 12000  37000
#> 16                       <NA> 25 - 44  11000  36000 21000  50000
#> 17                       <NA> 45 - 64  19000  91000 17000  75000
#> 18                       <NA>     65+  16000 118000  9000  66000
#> 19 Postgraduate qualification 15 - 24   <NA>   6000  <NA>   <NA>
#> 20                       <NA> 25 - 44   5000  86000  7000  60000
#> 21                       <NA> 45 - 64   6000  55000  6000  68000
#> 22                       <NA>     65+   <NA>  13000  <NA>  18000

# Make a tidy representation
cells <- tidy_table(x)
cells <- cells[!is.na(cells$chr), ] # Introduce 'holes' in the data
# Select a particular cell
cell <- cells[which(cells$row == 3 & cells$col == 3), ]
# Extend the selection downwards, stopping before the NA.
extend_S(cell, cells, boundary = ~ is.na(chr))

#> # A tibble: 3 x 3
#>     row   col chr  
#>   <int> <int> <chr>
#> 1     4     3 12000
#> 2     5     3 10000
#> 3     3     3 7000 

# Extend the selection right, up to and including the fifth column.
extend_E(cell, cells, boundary = ~ col == 5, include = TRUE)

#> # A tibble: 3 x 3
#>     row   col chr  
#>   <int> <int> <chr>
#> 1     3     4 27000
#> 2     3     5 <NA> 
#> 3     3     3 7000 

# Extend the selection beyond the existing cells
extend_E(cell, cells, 15)

#> # A tibble: 16 x 3
#>      row   col chr  
#>    <int> <int> <chr>
#>  1     3     4 27000
#>  2     3     5 <NA> 
#>  3     3     6 13000
#>  4     3     7 <NA> 
#>  5     3     8 <NA> 
#>  6     3     9 <NA> 
#>  7     3    10 <NA> 
#>  8     3    11 <NA> 
#>  9     3    12 <NA> 
#> 10     3    13 <NA> 
#> 11     3    14 <NA> 
#> 12     3    15 <NA> 
#> 13     3    16 <NA> 
#> 14     3    17 <NA> 
#> 15     3    18 <NA> 
#> 16     3     3 7000 

# This doesn't work inside formulas, because it would mean testing the
# boundary formula on every possible cell in the given direction

# NOT RUN {
extend_E(cell, cells, boundary = ~ col == 15)
# }
cell <- cells[which(cells$row == 7 & cells$col %in% 1:2), ]
extend_N(cell, cells, boundary = ~ !is.na(chr), edge = TRUE)

#> # A tibble: 8 x 3
#>     row   col chr        
#>   <int> <int> <chr>      
#> 1     4     1 <NA>       
#> 2     4     2 25 - 44    
#> 3     5     1 <NA>       
#> 4     5     2 45 - 64    
#> 5     6     1 <NA>       
#> 6     6     2 65+        
#> 7     7     1 Certificate
#> 8     7     2 15 - 24