Join a bag of data cells to some header, by proximity in a given direction,

A bag of data cells is a data frame with at least the columns 'row' and 'col', as well as any others that carry information about the cells, e.g. their values. Cells in a table are associated with header cells by proximity. Having collected header cells and data cells into separate data frames, 'join_header' and the related functions 'NNW', 'ABOVE', etc., join the values in the header cells to the data cells, choose the nearest header to each cell, in a given direction.

join_header(bag, header, direction, boundaries = NULL)

N(bag, header)

E(bag, header)

S(bag, header)

W(bag, header)

NNW(bag, header)

NNE(bag, header)

ENE(bag, header)

ESE(bag, header)

SSE(bag, header)

SSW(bag, header)

WSW(bag, header)

WNW(bag, header)

ABOVE(bag, header, boundaries = NULL)

LEFT(bag, header, boundaries = NULL)

BELOW(bag, header, boundaries = NULL)

RIGHT(bag, header, boundaries = NULL)

Arguments

bag

Data frame. A bag of data cells including at least the columns 'row' and 'column', which are numeric/integer vectors.

header

Data frame. A bag of data cells including at least the columns 'row' and 'column', which are numeric/integer vectors.

direction

Character vector length 1. A compass direction to search for the nearest header. See 'details'.

boundaries

Data frame. Only applies to the directions "ABOVE", "RIGHT", "BELOW" and "LEFT". A bag of cells in one row or one column, demarking boundaries within which to match headers with cells. For example, a boundary could be a bag of cells with borders on one side. This is useful when the nearest header might be the wrong header because it lies on the other side of a border.

Details

Headers are associated with data by proximity in a given direction. The directions are mapped to the points of the compass, where 'N' is north (up), 'E' is east (right), and so on. `join_header()` finds the nearest header to a given data cell in a given direction, and joins its value to the data cell. The most common directions to search are 'NNW' (for left-aligned headers at the top of the table) and 'WNW' for top-aligned headers at the side of the table. The difference between 'N' and 'ABOVE' (and similar pairs of directions) is that 'N' finds headers directly above the data cell, whereas 'ABOVE' matches the nearest header, whether above-left, above-right or directly above the data cell. This is useful for matching headers that are not aligned to the edge of the data cells that they refer to. There can be a tie in the directions 'ABOVE', 'BELOW', 'LEFT' and 'RIGHT' , causing NAs to be returned in the place of header values. The full list of available directions is 'N', 'E', 'S', 'W', 'NNW', 'NNE', 'ENE', 'ESE', 'SSE', 'SSW', 'WSW', 'WNW', 'ABOVE', 'BELOW', 'LEFT', 'RIGHT'. For convenience, these directions are provided as their own functions, wrapping the concept of 'join_header()'.

Functions

N: Join nearest header in the 'N' direction.

E: Join nearest header in the 'E' direction.

S: Join nearest header in the 'S' direction.

W: Join nearest header in the 'W' direction.

NNW: Join nearest header in the 'NNW' direction.

NNE: Join nearest header in the 'NNE' direction.

ENE: Join nearest header in the 'ENE' direction.

ESE: Join nearest header in the 'ESE' direction.

SSE: Join nearest header in the 'SSE' direction.

SSW: Join nearest header in the 'SSW' direction.

WSW: Join nearest header in the 'WSW' direction.

WNW: Join nearest header in the 'WNW' direction.

ABOVE: Join nearest header in the 'ABOVE' direction.

LEFT: Join nearest header in the 'LEFT' direction.

BELOW: Join nearest header in the 'BELOW' direction.

RIGHT: Join nearest header in the 'BELOW' direction.

Examples

library(dplyr)
# Load some pivoted data
(x <- purpose$`NNW WNW`)

#>                            X2      X3     X4     X5    X6     X7
#> 1                        <NA>    <NA> Female   <NA>  Male   <NA>
#> 2                        <NA>    <NA>  0 - 6 7 - 10 0 - 6 7 - 10
#> 3           Bachelor's degree 15 - 24   7000  27000  <NA>  13000
#> 4                        <NA> 25 - 44  12000 137000  9000  81000
#> 5                        <NA> 45 - 64  10000  64000  7000  66000
#> 6                        <NA>     65+   <NA>  18000  7000  17000
#> 7                 Certificate 15 - 24  29000 161000 30000 190000
#> 8                        <NA> 25 - 44  34000 179000 31000 219000
#> 9                        <NA> 45 - 64  30000 210000 23000 199000
#> 10                       <NA>     65+  12000  77000  8000 107000
#> 11                    Diploma 15 - 24   <NA>  14000  9000  11000
#> 12                       <NA> 25 - 44  10000  66000  8000  47000
#> 13                       <NA> 45 - 64   6000  68000  5000  58000
#> 14                       <NA>     65+   5000  41000  1000  34000
#> 15           No Qualification 15 - 24  10000  43000 12000  37000
#> 16                       <NA> 25 - 44  11000  36000 21000  50000
#> 17                       <NA> 45 - 64  19000  91000 17000  75000
#> 18                       <NA>     65+  16000 118000  9000  66000
#> 19 Postgraduate qualification 15 - 24   <NA>   6000  <NA>   <NA>
#> 20                       <NA> 25 - 44   5000  86000  7000  60000
#> 21                       <NA> 45 - 64   6000  55000  6000  68000
#> 22                       <NA>     65+   <NA>  13000  <NA>  18000

# Make a tidy representation
cells <- tidytable(x, FALSE, FALSE)
cells <- cells[!is.na(cells$character), ]
head(cells)

#>    row col                  character double integer logical
#> 3    3   1          Bachelor's degree     NA      NA      NA
#> 7    7   1                Certificate     NA      NA      NA
#> 11  11   1                    Diploma     NA      NA      NA
#> 15  15   1           No Qualification     NA      NA      NA
#> 19  19   1 Postgraduate qualification     NA      NA      NA
#> 25   3   2                    15 - 24     NA      NA      NA

# Select the cells containing the values
datacells <-
  cells %>%
  filter(row >= 3, col >= 3)
head(datacells)

#>   row col character double integer logical
#> 1   3   3      7000     NA      NA      NA
#> 2   4   3     12000     NA      NA      NA
#> 3   5   3     10000     NA      NA      NA
#> 4   7   3     29000     NA      NA      NA
#> 5   8   3     34000     NA      NA      NA
#> 6   9   3     30000     NA      NA      NA

# Select the row headers
row_headers <-
  cells %>%
  filter(col <= 2) %>%
  select(row, col, header = character) %>%
  split(.$col) # Separate each column of headers
row_headers

#> $`1`
#>   row col                     header
#> 1   3   1          Bachelor's degree
#> 2   7   1                Certificate
#> 3  11   1                    Diploma
#> 4  15   1           No Qualification
#> 5  19   1 Postgraduate qualification
#> 
#> $`2`
#>    row col  header
#> 6    3   2 15 - 24
#> 7    4   2 25 - 44
#> 8    5   2 45 - 64
#> 9    6   2     65+
#> 10   7   2 15 - 24
#> 11   8   2 25 - 44
#> 12   9   2 45 - 64
#> 13  10   2     65+
#> 14  11   2 15 - 24
#> 15  12   2 25 - 44
#> 16  13   2 45 - 64
#> 17  14   2     65+
#> 18  15   2 15 - 24
#> 19  16   2 25 - 44
#> 20  17   2 45 - 64
#> 21  18   2     65+
#> 22  19   2 15 - 24
#> 23  20   2 25 - 44
#> 24  21   2 45 - 64
#> 25  22   2     65+
#> 

# Select the column headers
col_headers <-
  cells %>%
  filter(row <= 2) %>%
  select(row, col, header = character) %>%
  split(.$row) # Separate each row of headers
col_headers

#> $`1`
#>   row col header
#> 1   1   3 Female
#> 4   1   5   Male
#> 
#> $`2`
#>   row col header
#> 2   2   3  0 - 6
#> 3   2   4 7 - 10
#> 5   2   5  0 - 6
#> 6   2   6 7 - 10
#> 

# From each data cell, search for the nearest one of each of the headers
datacells %>%
  NNW(col_headers$`1`) %>%
  N(col_headers$`2`) %>%
  WNW(row_headers$`1`) %>%
  W(row_headers$`2`)

#> # A tibble: 72 × 10
#>      row   col character double integer logical header.data header.header
#>    <int> <int>     <chr>  <dbl>   <int>   <lgl>       <chr>         <chr>
#> 1      3     3      7000     NA      NA      NA      Female         0 - 6
#> 2      4     3     12000     NA      NA      NA      Female         0 - 6
#> 3      5     3     10000     NA      NA      NA      Female         0 - 6
#> 4      7     3     29000     NA      NA      NA      Female         0 - 6
#> 5      8     3     34000     NA      NA      NA      Female         0 - 6
#> 6      9     3     30000     NA      NA      NA      Female         0 - 6
#> 7     10     3     12000     NA      NA      NA      Female         0 - 6
#> 8     12     3     10000     NA      NA      NA      Female         0 - 6
#> 9     13     3      6000     NA      NA      NA      Female         0 - 6
#> 10    14     3      5000     NA      NA      NA      Female         0 - 6
#> # ... with 62 more rows, and 2 more variables: header.data.data <chr>,
#> #   header.header.header <chr>