Pad a bag of cells with blanks

When source data is sparse (e.g. spreadsheets from the 'tidyxl' package), the output of tidy_table may also be sparse. To fill the gaps with blank cells, use pad. Internally, pad is used to ensure that functions like offset and extend behave intuitively, as though blank cells do actually exist.

pad(cells, rows = cells$row, cols = cells$col)

Arguments

cells
Data frame, the cells to pad

rows
Numeric vector, the rows to be padded (existing cells will never be discarded, e.g. if this is 0, and no gaps will be left even if this is outside the original range of rows)

cols
Numeric vector, the columns to be padded (existing cells will never be discarded, e.g. if this is 0, and no gaps will be left even if this is outside the original range of columns)

Examples

cells <-
  tidy_table(purpose$`NNW WNW`) %>%
  dplyr::filter(!is.na(chr)) # Introduce 'holes' in the data
# Select a region with gaps
(bag <- dplyr::filter(cells, row %in% 2:4, col %in% 1:2))

#> # A tibble: 3 x 3
#>     row   col chr              
#>   <int> <int> <chr>            
#> 1     3     1 Bachelor's degree
#> 2     3     2 15 - 24          
#> 3     4     2 25 - 44          

# Pad the gaps
pad(bag) # By default, the selection is squared-off to its width and height

#> # A tibble: 4 x 3
#>     row   col chr              
#>   <int> <int> <chr>            
#> 1     3     1 Bachelor's degree
#> 2     3     2 15 - 24          
#> 3     4     1 <NA>             
#> 4     4     2 25 - 44          

pad(bag, 0, 0) # Zeros are equivalent to the defaults

#> # A tibble: 4 x 3
#>     row   col chr              
#>   <dbl> <dbl> <chr>            
#> 1  3.00  1.00 Bachelor's degree
#> 2  3.00  2.00 15 - 24          
#> 3  4.00  1.00 <NA>             
#> 4  4.00  2.00 25 - 44          

pad(bag, 2:5, 1:3) # Add a row and a column

#> # A tibble: 12 x 3
#>      row   col chr              
#>    <int> <int> <chr>            
#>  1     2     1 <NA>             
#>  2     2     2 <NA>             
#>  3     2     3 <NA>             
#>  4     3     1 Bachelor's degree
#>  5     3     2 15 - 24          
#>  6     3     3 <NA>             
#>  7     4     1 <NA>             
#>  8     4     2 25 - 44          
#>  9     4     3 <NA>             
#> 10     5     1 <NA>             
#> 11     5     2 <NA>             
#> 12     5     3 <NA>             

pad(bag, 6, 4) # No gaps are left even when .rows or .cols are distant

#> # A tibble: 16 x 3
#>      row   col chr              
#>    <dbl> <dbl> <chr>            
#>  1  3.00  1.00 Bachelor's degree
#>  2  3.00  2.00 15 - 24          
#>  3  3.00  3.00 <NA>             
#>  4  3.00  4.00 <NA>             
#>  5  4.00  1.00 <NA>             
#>  6  4.00  2.00 25 - 44          
#>  7  4.00  3.00 <NA>             
#>  8  4.00  4.00 <NA>             
#>  9  5.00  1.00 <NA>             
#> 10  5.00  2.00 <NA>             
#> 11  5.00  3.00 <NA>             
#> 12  5.00  4.00 <NA>             
#> 13  6.00  1.00 <NA>             
#> 14  6.00  2.00 <NA>             
#> 15  6.00  3.00 <NA>             
#> 16  6.00  4.00 <NA>