Convert data frames into a tidy structure

Data frames represent data in a tabular structure. tidy_table takes the row and column position of each 'cell', and returns that information in a new data frame, alongside the content of each cell.

This makes certain tasks easier. For example, a pivot table with multi-row headers that has been imported into R as a data frame may be easier to un-pivot by converting it with tidy_table first.

For HTML tables, the content of each cell is returned as standalone HTML that can be further parsed with tools such as the rvest package. This is particularly useful when an HTML cell itself contains an HTML table, or contains both text and a URL, which must be extracted separately. If the HTML itself is poorly formatted, try passing it through the htmltidy package first.

This is an S3 generic.

tidy_table(x, rownames = FALSE, colnames = FALSE)

Arguments

x
A data.frame or an HTML document

rownames
Whether to include the column names in the output, Default: FALSE

colnames
Whether to include the row names in the output, Default: FALSE

Value

A data.frame with columns 'row' and 'col' (integer) giving the original position of the 'cells', and any relevant columns for cell values in their original types: 'chr', 'cplx', 'cplx', 'dbl', 'fctr', 'int', 'lgl', and 'list'. The columns 'fctr' is, like 'list', a list-column (each element is itself a list) to avoid factor levels clashing. For HTML tables, the column 'html' gives the HTML of the original cell.

Row and column names, when present and required, are treated as though they were cells in the table, and they appear in the 'chr' column.

Examples

tidy_table(Formaldehyde)

#> # A tibble: 12 x 3
#>      row   col    dbl
#>    <int> <int>  <dbl>
#>  1     1     1 0.100 
#>  2     2     1 0.300 
#>  3     3     1 0.500 
#>  4     4     1 0.600 
#>  5     5     1 0.700 
#>  6     6     1 0.900 
#>  7     1     2 0.0860
#>  8     2     2 0.269 
#>  9     3     2 0.446 
#> 10     4     2 0.538 
#> 11     5     2 0.626 
#> 12     6     2 0.782 

tidy_table(tidyr::nest(chickwts, -feed))

#> # A tibble: 12 x 4
#>      row   col fctr       list                 
#>    <int> <int> <list>     <list>               
#>  1     1     1 <fctr [1]> <NULL>               
#>  2     2     1 <fctr [1]> <NULL>               
#>  3     3     1 <fctr [1]> <NULL>               
#>  4     4     1 <fctr [1]> <NULL>               
#>  5     5     1 <fctr [1]> <NULL>               
#>  6     6     1 <fctr [1]> <NULL>               
#>  7     1     2 <NULL>     <data.frame [10 × 1]>
#>  8     2     2 <NULL>     <data.frame [12 × 1]>
#>  9     3     2 <NULL>     <data.frame [14 × 1]>
#> 10     4     2 <NULL>     <data.frame [12 × 1]>
#> 11     5     2 <NULL>     <data.frame [11 × 1]>
#> 12     6     2 <NULL>     <data.frame [12 × 1]>

tidy_table(Formaldehyde, colnames = TRUE)

#> # A tibble: 14 x 4
#>      row   col chr        dbl
#>    <int> <int> <chr>    <dbl>
#>  1     1     1 carb   NA     
#>  2     2     1 <NA>    0.100 
#>  3     3     1 <NA>    0.300 
#>  4     4     1 <NA>    0.500 
#>  5     5     1 <NA>    0.600 
#>  6     6     1 <NA>    0.700 
#>  7     7     1 <NA>    0.900 
#>  8     1     2 optden NA     
#>  9     2     2 <NA>    0.0860
#> 10     3     2 <NA>    0.269 
#> 11     4     2 <NA>    0.446 
#> 12     5     2 <NA>    0.538 
#> 13     6     2 <NA>    0.626 
#> 14     7     2 <NA>    0.782 

tidy_table(Formaldehyde, rownames = TRUE)

#> # A tibble: 18 x 4
#>      row   col chr       dbl
#>    <int> <int> <chr>   <dbl>
#>  1     1     1 1     NA     
#>  2     2     1 2     NA     
#>  3     3     1 3     NA     
#>  4     4     1 4     NA     
#>  5     5     1 5     NA     
#>  6     6     1 6     NA     
#>  7     1     2 <NA>   0.100 
#>  8     2     2 <NA>   0.300 
#>  9     3     2 <NA>   0.500 
#> 10     4     2 <NA>   0.600 
#> 11     5     2 <NA>   0.700 
#> 12     6     2 <NA>   0.900 
#> 13     1     3 <NA>   0.0860
#> 14     2     3 <NA>   0.269 
#> 15     3     3 <NA>   0.446 
#> 16     4     3 <NA>   0.538 
#> 17     5     3 <NA>   0.626 
#> 18     6     3 <NA>   0.782 

colspan <- system.file("extdata", "colspan.html", package = "unpivotr")
rowspan <- system.file("extdata", "rowspan.html", package = "unpivotr")
nested <- system.file("extdata", "nested.html", package = "unpivotr")
# NOT RUN {
browseURL(colspan)
browseURL(rowspan)
browseURL(nestedspan)
# }

tidy_table(xml2::read_html(colspan))

#> [[1]]
#> # A tibble: 4 x 3
#>     row   col html                                    
#>   <int> <int> <chr>                                   
#> 1     1     1 "<th colspan=\"2\">Header (1, 1:2)</th>"
#> 2     2     1 <td>cell (2, 1)</td>                    
#> 3     1     2 <NA>                                    
#> 4     2     2 <td>cell (2, 2)</td>                    
#> 

tidy_table(xml2::read_html(rowspan))

#> [[1]]
#> # A tibble: 4 x 3
#>     row   col html                                    
#>   <int> <int> <chr>                                   
#> 1     1     1 "<th rowspan=\"2\">Header (1:2, 1)</th>"
#> 2     2     1 <NA>                                    
#> 3     1     2 <th>Header (1, 2)</th>                  
#> 4     2     2 <td>cell (2, 2)</td>                    
#> 

tidy_table(xml2::read_html(nested))

#> [[1]]
#> # A tibble: 4 x 3
#>     row   col html                                                             
#>   <int> <int> <chr>                                                            
#> 1     1     1 <th>Header (1, 1)</th>                                           
#> 2     2     1 <td>cell (2, 1)</td>                                             
#> 3     1     2 <th>Header (1, 2)</th>                                           
#> 4     2     2 "<td>\n          <table>\n<tr>\n<th>Header (2, 2)(1, 1)</th>\n  …
#>