Convert data frames into a tidy structure

Data frames and matrices represents data in a tabular structure. tidytable takes the row and column position of each 'cell', and returns that information in a new data frame. This makes certain tasks easier. For example, a pivot table with multi-row headers that has been imported into R as a data frame may be easier to un-pivot by converting it with tidytable first.

This is an S3 generic.

tidytable(x, rownames = TRUE, colnames = TRUE)

Arguments

x

A data.frame or a matrix

rownames

Whether to include the column names in the output

colnames

Whether to include the row names in the output

Value

A data.frame with six columns: 'row' and 'col' (integer) giving the original position of the 'cells', and 'character', 'double', 'integer' and 'logical' giving the cell values in their original types. Row and column names are returned in the 'character' column, and, when present, offset the other cells by one row or column.

Examples

tidytable(Formaldehyde)

#>    row col character double integer logical
#> 1    2   2      <NA>  0.100      NA      NA
#> 2    3   2      <NA>  0.300      NA      NA
#> 3    4   2      <NA>  0.500      NA      NA
#> 4    5   2      <NA>  0.600      NA      NA
#> 5    6   2      <NA>  0.700      NA      NA
#> 6    7   2      <NA>  0.900      NA      NA
#> 7    2   3      <NA>  0.086      NA      NA
#> 8    3   3      <NA>  0.269      NA      NA
#> 9    4   3      <NA>  0.446      NA      NA
#> 10   5   3      <NA>  0.538      NA      NA
#> 11   6   3      <NA>  0.626      NA      NA
#> 12   7   3      <NA>  0.782      NA      NA
#> 13   2   1         1     NA      NA      NA
#> 14   3   1         2     NA      NA      NA
#> 15   4   1         3     NA      NA      NA
#> 16   5   1         4     NA      NA      NA
#> 17   6   1         5     NA      NA      NA
#> 18   7   1         6     NA      NA      NA
#> 19   1   2      carb     NA      NA      NA
#> 20   1   3    optden     NA      NA      NA

tidytable(as.matrix(Formaldehyde))

#>    row col character double integer logical
#> 1    2   2      <NA>  0.100      NA      NA
#> 2    3   2      <NA>  0.300      NA      NA
#> 3    4   2      <NA>  0.500      NA      NA
#> 4    5   2      <NA>  0.600      NA      NA
#> 5    6   2      <NA>  0.700      NA      NA
#> 6    7   2      <NA>  0.900      NA      NA
#> 7    2   3      <NA>  0.086      NA      NA
#> 8    3   3      <NA>  0.269      NA      NA
#> 9    4   3      <NA>  0.446      NA      NA
#> 10   5   3      <NA>  0.538      NA      NA
#> 11   6   3      <NA>  0.626      NA      NA
#> 12   7   3      <NA>  0.782      NA      NA
#> 13   2   1         1     NA      NA      NA
#> 14   3   1         2     NA      NA      NA
#> 15   4   1         3     NA      NA      NA
#> 16   5   1         4     NA      NA      NA
#> 17   6   1         5     NA      NA      NA
#> 18   7   1         6     NA      NA      NA
#> 19   1   2      carb     NA      NA      NA
#> 20   1   3    optden     NA      NA      NA

tidytable(Formaldehyde, colnames = FALSE)

#>    row col character double integer logical
#> 1    1   2      <NA>  0.100      NA      NA
#> 2    2   2      <NA>  0.300      NA      NA
#> 3    3   2      <NA>  0.500      NA      NA
#> 4    4   2      <NA>  0.600      NA      NA
#> 5    5   2      <NA>  0.700      NA      NA
#> 6    6   2      <NA>  0.900      NA      NA
#> 7    1   3      <NA>  0.086      NA      NA
#> 8    2   3      <NA>  0.269      NA      NA
#> 9    3   3      <NA>  0.446      NA      NA
#> 10   4   3      <NA>  0.538      NA      NA
#> 11   5   3      <NA>  0.626      NA      NA
#> 12   6   3      <NA>  0.782      NA      NA
#> 13   1   1         1     NA      NA      NA
#> 14   2   1         2     NA      NA      NA
#> 15   3   1         3     NA      NA      NA
#> 16   4   1         4     NA      NA      NA
#> 17   5   1         5     NA      NA      NA
#> 18   6   1         6     NA      NA      NA

tidytable(Formaldehyde, rownames = FALSE)

#>    row col character double integer logical
#> 1    2   1      <NA>  0.100      NA      NA
#> 2    3   1      <NA>  0.300      NA      NA
#> 3    4   1      <NA>  0.500      NA      NA
#> 4    5   1      <NA>  0.600      NA      NA
#> 5    6   1      <NA>  0.700      NA      NA
#> 6    7   1      <NA>  0.900      NA      NA
#> 7    2   2      <NA>  0.086      NA      NA
#> 8    3   2      <NA>  0.269      NA      NA
#> 9    4   2      <NA>  0.446      NA      NA
#> 10   5   2      <NA>  0.538      NA      NA
#> 11   6   2      <NA>  0.626      NA      NA
#> 12   7   2      <NA>  0.782      NA      NA
#> 13   1   1      carb     NA      NA      NA
#> 14   1   2    optden     NA      NA      NA