Pad a bag of cells with blanks

When source data is sparse (e.g. spreadsheets from the 'tidyxl' package), the output of tidytable may also be sparse. To fill the gaps with blank cells, use pad. Internally, pad is used to ensure that functions like offset and extend behave intuitively, as though blank cells do actually exist.

pad(cells, rows = cells$row, cols = cells$col)

Arguments

cells

Data frame, the cells to pad

rows

Numeric vector, the rows to be padded (existing cells will never be discarded, e.g. if this is 0, and no gaps will be left even if this is outside the original range of rows)

cols

Numeric vector, the columns to be padded (existing cells will never be discarded, e.g. if this is 0, and no gaps will be left even if this is outside the original range of columns)

Examples

cells <-
  tidytable(purpose$`NNW WNW`, FALSE, FALSE) %>%
  dplyr::filter(!is.na(character)) # Introduce 'holes' in the data
# Select a region with gaps
(bag <- dplyr::filter(cells, row %in% 2:4, col %in% 1:2))

#>   row col         character double integer logical
#> 1   3   1 Bachelor's degree     NA      NA      NA
#> 2   3   2           15 - 24     NA      NA      NA
#> 3   4   2           25 - 44     NA      NA      NA

# Pad the gaps
pad(bag) # By default, the selection is squared-off to its width and height

#> # A tibble: 4 × 6
#>     row   col         character double integer logical
#>   <int> <int>             <chr>  <dbl>   <int>   <lgl>
#> 1     3     1 Bachelor's degree     NA      NA      NA
#> 2     3     2           15 - 24     NA      NA      NA
#> 3     4     1              <NA>     NA      NA      NA
#> 4     4     2           25 - 44     NA      NA      NA

pad(bag, 0, 0) # Zeros are equivalent to the defaults

#> # A tibble: 4 × 6
#>     row   col         character double integer logical
#>   <dbl> <dbl>             <chr>  <dbl>   <int>   <lgl>
#> 1     3     1 Bachelor's degree     NA      NA      NA
#> 2     3     2           15 - 24     NA      NA      NA
#> 3     4     1              <NA>     NA      NA      NA
#> 4     4     2           25 - 44     NA      NA      NA

pad(bag, 2:5, 1:3) # Add a row and a column

#> # A tibble: 12 × 6
#>      row   col         character double integer logical
#>    <int> <int>             <chr>  <dbl>   <int>   <lgl>
#> 1      2     1              <NA>     NA      NA      NA
#> 2      2     2              <NA>     NA      NA      NA
#> 3      2     3              <NA>     NA      NA      NA
#> 4      3     1 Bachelor's degree     NA      NA      NA
#> 5      3     2           15 - 24     NA      NA      NA
#> 6      3     3              <NA>     NA      NA      NA
#> 7      4     1              <NA>     NA      NA      NA
#> 8      4     2           25 - 44     NA      NA      NA
#> 9      4     3              <NA>     NA      NA      NA
#> 10     5     1              <NA>     NA      NA      NA
#> 11     5     2              <NA>     NA      NA      NA
#> 12     5     3              <NA>     NA      NA      NA

pad(bag, 6, 4) # No gaps are left even when .rows or .cols are distant

#> # A tibble: 16 × 6
#>      row   col         character double integer logical
#>    <dbl> <dbl>             <chr>  <dbl>   <int>   <lgl>
#> 1      3     1 Bachelor's degree     NA      NA      NA
#> 2      3     2           15 - 24     NA      NA      NA
#> 3      3     3              <NA>     NA      NA      NA
#> 4      3     4              <NA>     NA      NA      NA
#> 5      4     1              <NA>     NA      NA      NA
#> 6      4     2           25 - 44     NA      NA      NA
#> 7      4     3              <NA>     NA      NA      NA
#> 8      4     4              <NA>     NA      NA      NA
#> 9      5     1              <NA>     NA      NA      NA
#> 10     5     2              <NA>     NA      NA      NA
#> 11     5     3              <NA>     NA      NA      NA
#> 12     5     4              <NA>     NA      NA      NA
#> 13     6     1              <NA>     NA      NA      NA
#> 14     6     2              <NA>     NA      NA      NA
#> 15     6     3              <NA>     NA      NA      NA
#> 16     6     4              <NA>     NA      NA      NA