Divide a grid of cells into partitions containing individual tables
Source: R/partition.R

Given the positions of corner cells that mark individual tables in a single spreadsheet, partion() works out which cells belong to which tables, and gives them an ID. This can be used with dplyr::group_by() or tidyr::nest() to operate on each single table within a spreadsheet of several tables.

Corners must all be in the top-left, or bottom-left, or top-right, or bottom-right of each table.

partition() partitions along both dimensions (rows and columns) at once. partition_dim() partitions along one dimension at a time.

partition(cells, corners, position = "top_left",
  partition_name = "partition")

partition_dim(positions, cutpoints, bound = "upper")

Arguments

cells
Data frame or tbl, the cells to be partitioned, from tidy_table() or tidyxl::xlsx_cells().

corners
a subset of cells, being the corners of individual tables.

position
Character, the position of the corner cells relative to their tables, one of "top-left" (default), "top-right", "bottom-left", "bottom-right". position or the column position), which are to be grouped between cutpoints.

partition_name
Character vector length 1, what to call the new column that will be created to identify the partitions. Default: "partition".

positions
Integer vector, the positions of cells (either the row position or the column position), which are to be grouped between cutpoints.

cutpoints
Integer vector. The positions will be separated into groups either side of each cutpoint.

bound
One of "upper" or "lower", controls whether cells that lie on a cutpoint are should be grouped with cells below or above the cutpoint. For example, if column 5 is a cutpoint, and a cell is in column 5, "lower" would group it with cells in columns 1 to 4, whereas "upper" would group it with cells in columns 6 to 10. This is so that you can use cells at the bottom or the right-hand side of a table as the cutpoints (either of which would be 'upper' bounds because row and column numbers count from 1 in the top-left row and column). When "upper", any cell_positions above the first cutpoint will be in group 0; when "lower", any cell_positions below the final cutpoint will be 0.

Value

partition_dim() returns an integer vector, numbering the groups of cells. Group 0 represents the cells above the first cutpoint (when bound = "upper"), or below the first cutpoint (when bound = "lower"). The other groups are numbered from 1, where group 1 is adjacent to group 0.

partition_dim() returns an integer vector, numbering the groups of cells. Group 0 represents the cells above the first cutpoint (when bound = "upper"), or below the first cutpoint (when bound = "lower"). The other groups are numbered from 1, where group 1 is adjacent to group 0. Divide a grid of cells into chunks along both dimensions

Functions

partition_dim: Divide a grid of cells into chunks along one dimension

Examples

# The `purpose` dataset, represented in four summary tables
multiples <- purpose$small_multiples
rectify(multiples, character, numeric)

#> # A tibble: 14 x 6
#>    `row/col` `1(A)`                     `2(B)` `3(C)` `4(D)`            `5(E)`
#>        <int> <chr>                      <chr>  <lgl>  <chr>             <chr> 
#>  1         1 Postgraduate qualification <NA>   NA     Bachelor's degree <NA>  
#>  2         2 Sex                        Value  NA     Sex               Value 
#>  3         3 Female                     171000 NA     Female            275000
#>  4         4 Male                       159000 NA     Male              2e+05 
#>  5         5 <NA>                       <NA>   NA     <NA>              <NA>  
#>  6         6 Diploma                    <NA>   NA     Certificate       <NA>  
#>  7         7 Sex                        Value  NA     Sex               Value 
#>  8         8 Female                     210000 NA     Female            732000
#>  9         9 Male                       173000 NA     Male              807000
#> 10        10 <NA>                       <NA>   NA     <NA>              <NA>  
#> 11        11 No Qualification           <NA>   NA     <NA>              <NA>  
#> 12        12 Sex                        Value  NA     <NA>              <NA>  
#> 13        13 Female                     344000 NA     <NA>              <NA>  
#> 14        14 Male                       287000 NA     <NA>              <NA>  

# The same thing in its raw 'melted' form that can be filtered
multiples

#>    row col data_type                  character numeric
#> 1    1   1 character Postgraduate qualification      NA
#> 2    1   2     blank                       <NA>      NA
#> 3    1   4 character          Bachelor's degree      NA
#> 4    1   5     blank                       <NA>      NA
#> 5    2   1 character                        Sex      NA
#> 6    2   2 character                      Value      NA
#> 7    2   4 character                        Sex      NA
#> 8    2   5 character                      Value      NA
#> 9    3   1 character                     Female      NA
#> 10   3   2   numeric                       <NA>  171000
#> 11   3   4 character                     Female      NA
#> 12   3   5   numeric                       <NA>  275000
#> 13   4   1 character                       Male      NA
#> 14   4   2   numeric                       <NA>  159000
#> 15   4   4 character                       Male      NA
#> 16   4   5   numeric                       <NA>  200000
#> 17   6   1 character                    Diploma      NA
#> 18   6   2     blank                       <NA>      NA
#> 19   6   4 character                Certificate      NA
#> 20   6   5     blank                       <NA>      NA
#> 21   7   1 character                        Sex      NA
#> 22   7   2 character                      Value      NA
#> 23   7   4 character                        Sex      NA
#> 24   7   5 character                      Value      NA
#> 25   8   1 character                     Female      NA
#> 26   8   2   numeric                       <NA>  210000
#> 27   8   4 character                     Female      NA
#> 28   8   5   numeric                       <NA>  732000
#> 29   9   1 character                       Male      NA
#> 30   9   2   numeric                       <NA>  173000
#> 31   9   4 character                       Male      NA
#> 32   9   5   numeric                       <NA>  807000
#> 33  11   1 character           No Qualification      NA
#> 34  11   2     blank                       <NA>      NA
#> 35  12   1 character                        Sex      NA
#> 36  12   2 character                      Value      NA
#> 37  13   1 character                     Female      NA
#> 38  13   2   numeric                       <NA>  344000
#> 39  14   1 character                       Male      NA
#> 40  14   2   numeric                       <NA>  287000

# First, find the cells that mark a corner of each table
corners <-
  dplyr::filter(multiples,
                !is.na(character),
                !(character %in% c("Sex", "Value", "Female", "Male")))

# Then find out which cells fall into which partition
partition(multiples, corners)

#>    row col data_type                  character numeric partition
#> 1    1   1 character Postgraduate qualification      NA         1
#> 2    1   2     blank                       <NA>      NA         1
#> 3    1   4 character          Bachelor's degree      NA         2
#> 4    1   5     blank                       <NA>      NA         2
#> 5    2   1 character                        Sex      NA         1
#> 6    2   2 character                      Value      NA         1
#> 7    2   4 character                        Sex      NA         2
#> 8    2   5 character                      Value      NA         2
#> 9    3   1 character                     Female      NA         1
#> 10   3   2   numeric                       <NA>  171000         1
#> 11   3   4 character                     Female      NA         2
#> 12   3   5   numeric                       <NA>  275000         2
#> 13   4   1 character                       Male      NA         1
#> 14   4   2   numeric                       <NA>  159000         1
#> 15   4   4 character                       Male      NA         2
#> 16   4   5   numeric                       <NA>  200000         2
#> 17   6   1 character                    Diploma      NA         3
#> 18   6   2     blank                       <NA>      NA         3
#> 19   6   4 character                Certificate      NA         4
#> 20   6   5     blank                       <NA>      NA         4
#> 21   7   1 character                        Sex      NA         3
#> 22   7   2 character                      Value      NA         3
#> 23   7   4 character                        Sex      NA         4
#> 24   7   5 character                      Value      NA         4
#> 25   8   1 character                     Female      NA         3
#> 26   8   2   numeric                       <NA>  210000         3
#> 27   8   4 character                     Female      NA         4
#> 28   8   5   numeric                       <NA>  732000         4
#> 29   9   1 character                       Male      NA         3
#> 30   9   2   numeric                       <NA>  173000         3
#> 31   9   4 character                       Male      NA         4
#> 32   9   5   numeric                       <NA>  807000         4
#> 33  11   1 character           No Qualification      NA         5
#> 34  11   2     blank                       <NA>      NA         5
#> 35  12   1 character                        Sex      NA         5
#> 36  12   2 character                      Value      NA         5
#> 37  13   1 character                     Female      NA         5
#> 38  13   2   numeric                       <NA>  344000         5
#> 39  14   1 character                       Male      NA         5
#> 40  14   2   numeric                       <NA>  287000         5

# It's more obvious if you also split() by the new 'partition' column
x <- partition(multiples, corners)
split(x, x$partition)

#> $`1`
#>    row col data_type                  character numeric partition
#> 1    1   1 character Postgraduate qualification      NA         1
#> 2    1   2     blank                       <NA>      NA         1
#> 5    2   1 character                        Sex      NA         1
#> 6    2   2 character                      Value      NA         1
#> 9    3   1 character                     Female      NA         1
#> 10   3   2   numeric                       <NA>  171000         1
#> 13   4   1 character                       Male      NA         1
#> 14   4   2   numeric                       <NA>  159000         1
#> 
#> $`2`
#>    row col data_type         character numeric partition
#> 3    1   4 character Bachelor's degree      NA         2
#> 4    1   5     blank              <NA>      NA         2
#> 7    2   4 character               Sex      NA         2
#> 8    2   5 character             Value      NA         2
#> 11   3   4 character            Female      NA         2
#> 12   3   5   numeric              <NA>  275000         2
#> 15   4   4 character              Male      NA         2
#> 16   4   5   numeric              <NA>  200000         2
#> 
#> $`3`
#>    row col data_type character numeric partition
#> 17   6   1 character   Diploma      NA         3
#> 18   6   2     blank      <NA>      NA         3
#> 21   7   1 character       Sex      NA         3
#> 22   7   2 character     Value      NA         3
#> 25   8   1 character    Female      NA         3
#> 26   8   2   numeric      <NA>  210000         3
#> 29   9   1 character      Male      NA         3
#> 30   9   2   numeric      <NA>  173000         3
#> 
#> $`4`
#>    row col data_type   character numeric partition
#> 19   6   4 character Certificate      NA         4
#> 20   6   5     blank        <NA>      NA         4
#> 23   7   4 character         Sex      NA         4
#> 24   7   5 character       Value      NA         4
#> 27   8   4 character      Female      NA         4
#> 28   8   5   numeric        <NA>  732000         4
#> 31   9   4 character        Male      NA         4
#> 32   9   5   numeric        <NA>  807000         4
#> 
#> $`5`
#>    row col data_type        character numeric partition
#> 33  11   1 character No Qualification      NA         5
#> 34  11   2     blank             <NA>      NA         5
#> 35  12   1 character              Sex      NA         5
#> 36  12   2 character            Value      NA         5
#> 37  13   1 character           Female      NA         5
#> 38  13   2   numeric             <NA>  344000         5
#> 39  14   1 character             Male      NA         5
#> 40  14   2   numeric             <NA>  287000         5
#> 

# An alternative to split() is tidyr::nest()
tidyr::nest(x, -partition)

#> # A tibble: 5 x 2
#>   partition data                
#>       <int> <list>              
#> 1         1 <data.frame [8 × 5]>
#> 2         2 <data.frame [8 × 5]>
#> 3         3 <data.frame [8 × 5]>
#> 4         4 <data.frame [8 × 5]>
#> 5         5 <data.frame [8 × 5]>

# You can also use bottom-left corners (or top-right or bottom-right)
bl_corners <- dplyr::filter(multiples, character == "Yes")
y <- partition(multiples, bl_corners, position = "bottom_left")
split(y, y$partition)

#> named list()

# If `partition_name` is already the name of a column in `cells`, then it
# will be silently overwritten
multiples$important_column <- "Will be overwritten"
partition(multiples, corners, partition_name = "important_column")

#>    row col data_type                  character numeric important_column
#> 1    1   1 character Postgraduate qualification      NA                1
#> 2    1   2     blank                       <NA>      NA                1
#> 3    1   4 character          Bachelor's degree      NA                2
#> 4    1   5     blank                       <NA>      NA                2
#> 5    2   1 character                        Sex      NA                1
#> 6    2   2 character                      Value      NA                1
#> 7    2   4 character                        Sex      NA                2
#> 8    2   5 character                      Value      NA                2
#> 9    3   1 character                     Female      NA                1
#> 10   3   2   numeric                       <NA>  171000                1
#> 11   3   4 character                     Female      NA                2
#> 12   3   5   numeric                       <NA>  275000                2
#> 13   4   1 character                       Male      NA                1
#> 14   4   2   numeric                       <NA>  159000                1
#> 15   4   4 character                       Male      NA                2
#> 16   4   5   numeric                       <NA>  200000                2
#> 17   6   1 character                    Diploma      NA                3
#> 18   6   2     blank                       <NA>      NA                3
#> 19   6   4 character                Certificate      NA                4
#> 20   6   5     blank                       <NA>      NA                4
#> 21   7   1 character                        Sex      NA                3
#> 22   7   2 character                      Value      NA                3
#> 23   7   4 character                        Sex      NA                4
#> 24   7   5 character                      Value      NA                4
#> 25   8   1 character                     Female      NA                3
#> 26   8   2   numeric                       <NA>  210000                3
#> 27   8   4 character                     Female      NA                4
#> 28   8   5   numeric                       <NA>  732000                4
#> 29   9   1 character                       Male      NA                3
#> 30   9   2   numeric                       <NA>  173000                3
#> 31   9   4 character                       Male      NA                4
#> 32   9   5   numeric                       <NA>  807000                4
#> 33  11   1 character           No Qualification      NA                5
#> 34  11   2     blank                       <NA>      NA                5
#> 35  12   1 character                        Sex      NA                5
#> 36  12   2 character                      Value      NA                5
#> 37  13   1 character                     Female      NA                5
#> 38  13   2   numeric                       <NA>  344000                5
#> 39  14   1 character                       Male      NA                5
#> 40  14   2   numeric                       <NA>  287000                5

# Given a set of cells in rows 1 to 10, partition them at the 3rd, 5th and 7th
# rows. Row 3 is the top row of group 1.
partition_dim(1:10, c(3, 5, 7))

#>  [1] 0 0 1 1 2 2 3 3 3 3

# Given a set of cells in columns 1 to 10, partition them at the 3rd, 5th and
# 7th column. Column 3 is the left-most column of group 1.
# This example is exactly the same as the previous one, to show that the
# function works the same way on columns as rows.
partition_dim(1:10, c(3, 5, 7))

#>  [1] 0 0 1 1 2 2 3 3 3 3

# Given a set of cells in rows 1 to 10, partition them at the 3rd, 5th and
# 7th rows. Row 7 is the bottom row of group 1
partition_dim(1:10, c(3, 5, 7), bound = "lower")

#>  [1] 3 3 3 2 2 1 1 0 0 0

# When the first row of cells is on the first cutpoint, there is no group 0.
partition_dim(1:10, c(1, 10))

#>  [1] 1 1 1 1 1 1 1 1 1 2

partition_dim(1:10, c(1, 10), bound = "lower")

#>  [1] 2 1 1 1 1 1 1 1 1 1

# Non-integer row/column numbers and cutpoints can be used, even though they
# make no sense in the context of partioning grids of cells.
partition_dim(1:10 - .5, c(3, 5, 7))

#>  [1] 0 0 0 1 1 2 2 3 3 3

partition_dim(1:10, c(3, 5, 7) + 1.5)

#>  [1] 0 0 0 0 1 1 2 2 3 3